Crate signalk

Signal K

signalk is a collections of types to serialize and deserialize the signal-k protocol.

Re-exports

pub use definitions::V1Attr;

pub use definitions::V1CommonValueFields;

pub use definitions::V1DefSource;

pub use definitions::V1Meta;

pub use definitions::V1MetaZone;

pub use definitions::V1NumberValue;

pub use delta::V1DeltaFormat;

pub use delta::V1UpdateMeta;

pub use delta::V1UpdateType;

pub use delta::V1UpdateValue;

pub use delta::V1UpdateValueType;

pub use discovery::V1Discovery;

pub use discovery::V1DiscoveryEndpoint;

pub use discovery::V1DiscoveryServer;

pub use electrical::V1ACBus;

pub use electrical::V1Electrical;

pub use electrical::V1ElectricalACQualities;

pub use electrical::V1ElectricalIdentity;

pub use environment::V1Environment;

pub use environment::V1EnvironmentCurrent;

pub use environment::V1EnvironmentCurrentValue;

pub use environment::V1EnvironmentDepth;

pub use environment::V1EnvironmentInside;

pub use environment::V1EnvironmentTime;

pub use full::V1FullFormat;

pub use hello::V1Hello;

pub use navigation::V1Navigation;

pub use navigation::V1PositionType;

pub use navigation::V1PositionValue;

pub use notification::V1Notification;

pub use notification::V1NotificationValue;

pub use propulsion::V1Propulsion;

pub use put::V1Put;

pub use put::V1PutValue;

pub use sources::V1Source;

pub use sources::V1SourceProperty;

pub use sources::V1Sources;

pub use subscribe::V1Subscribe;

pub use subscribe::V1Subscription;

pub use unsubscribe::V1Unsubscribe;

pub use unsubscribe::V1Unsubscription;

pub use vessel::V1Vessel;

Modules

communication

definitions

delta

discovery

electrical

environment

full

hello

navigation

notification

propulsion

put

sources

subscribe

unsubscribe

vessel

Structs

Storage

Enums

SignalKGetError

SignalKStreamMessage

Type for messages that can be received over the signal-k stream