signalk::full

Trait Updatable

pub trait Updatable {
    // Required methods
    fn apply_update(&mut self, update: &V1UpdateType);
    fn id(&self) -> String;
    fn type_name(&self) -> String;
}

These items can be updated by a V1UpdateType

Given the GetContext Trait one can fine a matching object to apply the update into from the V1DeltaFormat’s context

Required Methods

fn apply_update(&mut self, update: &V1UpdateType)

This will change the found object to have a a new value

fn id(&self) -> String

This will return the ID if possible for the object

At the moment this is mainly to make sure we can easily verify the GetContext functionality

fn type_name(&self) -> String

This will return the “type” of the context

At the moment this is mainly to make sure we can easily verify the GetContext functionality